csvファイルを追加するか初期化してから追加するか選択可能に

+++ app/src/routers/router.ts
@@ -285,12 +285,13 @@ router.post('/admin/csv/formHader', csrfProtection, async (req: Request, res: Re
 
 
     const startTimer = performance.now();
-    /* 初期化 */
-    if (await tagApplicationService.isExistTable()) await tagApplicationService.deleteAll();
-    await bookApplicationService.deleteBooks();
-    await publisherApplicationService.deletePublishers();
-    await authorApplicationService.deleteAuthors();
-
+    if (req.body.initData === 'true') {
+      /* 初期化 */
+      if (await tagApplicationService.isExistTable()) await tagApplicationService.deleteAll();
+      await bookApplicationService.deleteBooks();
+      await publisherApplicationService.deletePublishers();
+      await authorApplicationService.deleteAuthors();
+    }
     const csvLengh = csv.length;
 
     const booksPromise = [];
